bgpv2: move gobgp peer apis and unit tests to separate file
Moving add, remove and update neighbor APIs to separate file as
server.go is getting quite large with lot of functionality. Also,
splitting unit test file and moving peer tests to peer_test.go.
